Design test Linux to load linux image as a payload

+++ platform/platform.h
@@ -39,7 +39,7 @@ void platform_purge_core (void); // uarch state only
 void platform_init (void);
 void platform_core_init (void);
 
-#define platform_get_device_tree_addr() ((uintptr_t) (UNTRUSTED_ENTRY + PAYLOAD_MAXLEN))
+#define platform_get_device_tree_addr() ((uintptr_t) (((uint64_t) UNTRUSTED_ENTRY) + ((uint64_t) PAYLOAD_MAXLEN)))
 void platform_init_device_tree(void);
 
 void platform_interrupt_other_cores (void);
